Refactor document opening
