+++ studio/LonaStudio/Documents/JSONDocument.swift
@@ -41,7 +41,13 @@ class JSONDocument: NSDocument {
     var content: Content?
 
     override func makeWindowControllers() {
-        WorkspaceWindowController.create(andAttachTo: self)
+        DocumentController.shared.createOrFindWorkspaceWindowController(for: self)
+    }
+
+    override func showWindows() {
+        DocumentController.shared.createOrFindWorkspaceWindowController(for: self)
+
+        super.showWindows()
     }
 
     override func data(ofType typeName: String) throws -> Data {
