+++ studio/LonaStudio/Documents/MarkdownDocument.swift
@@ -37,7 +37,13 @@ class MarkdownDocument: NSDocument {
     }
 
     override func makeWindowControllers() {
-        WorkspaceWindowController.create(andAttachTo: self)
+        DocumentController.shared.createOrFindWorkspaceWindowController(for: self)
+    }
+
+    override func showWindows() {
+        DocumentController.shared.createOrFindWorkspaceWindowController(for: self)
+
+        super.showWindows()
     }
 
     override func data(ofType typeName: String) throws -> Data {
